implement InviteTeamForm
